+++ views/templates/hook/index.php
@@ -1,3 +1,14 @@
 <?php
+
+/**
+ * 2009-2026 Tecnoacquisti.com
+ *
+ * Directory access protection.
+ *
+ * @author    Arte e Informatica <helpdesk@tecnoacquisti.com>
+ * @copyright 2009-2026 Arte e Informatica
+ * @license   One Paid Licence By WebSite Using This Module. No Rent. No Sell. No Share.
+ */
+
 header('Location: ../../../');
 exit;